requirements.txt
# Hugging Face Spaces / pip install -r requirements.txt
streamlit>=1.29.0
pandas>=2.1.0
plotly>=6.1.1
google-genai>=1.0.0
datasets>=2.16.0
python-dotenv>=1.0.0
kaleido>=1.0
